drivers: adc: adc_context: allow using adc_context in shared module
Allow declaring the struct adc_context and options in a common header
to be nested in a common struct defined in that header. This allows
multiple implementations in case of RTIO based and "legacy" drivers
to share a common implementation of the adc_context.
